feat(injector): support asynchronous invoke()
Injector.invoke accepts optional 3rd parameter, a callback to notify that
the invoked function has completed.

If invoked with a callback, the context of the invoked function is supplied
with method `async()` (as in gruntjs), which returns a method for invoking
the supplied callback with a single value indicating the result of the method.

This could potentially be used to support asynchronous plugins in Karma, with
some work.